requirements.txt
crewai>=0.102.0
openai>=1.30.0
paramiko>=3.4.0
requests>=2.31.0
jinja2>=3.1.0
pyyaml>=6.0.1
click>=8.1.0
tenacity>=8.2.0
python-dotenv>=1.0.0
jsonpath-ng>=1.6.0
pandas>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
